Refactor and cleanup the TC UI Kit / Connect 2
Better organization of files, mirror structure of Bootstrap 4;
Remove dependency of Bourbon - this is a pretty big dependency/library, and I think we should avoid making the TC UI Kit dependent on third party libraries at this point. The goal is to allow the community to download the TC UI "Styles" folder and use it locally to develop static HTML/SCSS code, as well as translate this into a final React components;
Fixed the style of the "buttons";
Added forms - there are some changes pending there;
Split Mixins into a import file and sub-directory to make it easy to contain future expansion;
Started refactoring on the variable names - instead of using "label-large", we now should use "label-lg", "label-md", "label-sm", similarly to the way we use it in buttons. This is to keep the consistent pattern of expressing size through 2 letter code, which is self explanatory;
Removed Proxima Nova as typeface from our typography
Numerous small fixes
Should not break things; if we see breaking - let's fix the use of variable names.

In general - we don't have a good consistent use of variables, calculations, mixins. Let's try to straighten things going forward